ui: split redux boilerplate into Redux/store.js
This will allow us to import the store object from within container
components, so that we can dispatch actions to it.
